Create code
To use this code, you'll need to set up a Twilio account and purchase a phone number capable of receiving SMS messages. Then, you need to configure the webhook URL for incoming SMS messages to point to the route /sms on your server where this Flask app is hosted.

Please note that this script is a simplified example for educational purposes and may require additional configurations and security measures for production use. Additionally, you'll need to install the flask and twilio packages using pip.